requirements.txt
aioodbc==0.3.3
asgiref==3.3.4
async-lru==1.0.2
click==8.0.1
fastapi==0.65.2
h11==0.12.0
httptools==0.2.0
mmh3==3.0.0
numpy==1.21.1
pandas==1.2.4
pathlib==1.0.1
pydantic==1.8.2
pyodbc==4.0.30
python-dotenv==0.17.1
PyYAML==5.4.1
requests==2.25.1
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
#uvloop==0.15.2
watchgod==0.7
websockets==9.1
numba==0.56.0
xlrd==2.0.1
openpyxl==3.0.10
SQLAlchemy==1.4.41